Fix: delete duplications
