+++ src/main/app/client/gaServiceUrls.ts
@@ -2,3 +2,4 @@ export const GA_SERVICE_CASES_URL = '/cases/';
 export const GA_SERVICE_SUBMIT_EVENT = `${GA_SERVICE_CASES_URL}:caseId/citizen/:submitterId/event`;
 export const GA_FEES_PAYMENT_URL= '/fees/case/:claimId/payment';
 export const GA_FEES_PAYMENT_STATUS_URL= '/fees/case/:claimId/payment/:paymentReference/status';
+
