--- src/main/app/client/generalApplicationUrls.ts
@@ -1,1 +1,0 @@
-export const CIVIL_GENERAL_APPLICATIONS_URL = '/cases/';
